Every module has its own controller. The router.php has been wrote
